refactor: Twodsix only supports Foundry V9
BREAKING CHANGE: Twodsix only supports Foundry V9

After a couple of days (and nights) of hard work by @xdy, @jonepatr and @marvin9257 Twodsix now switches to only supporting Foundry V9 for all future versions.
As this is a breaking change we'll take the opportunity to go to version 1.0.0

This does *not* mean the end of development for Twodsix. Quite the contrary. :)