feat: make token attributes changable (#957)
This addition to the system makes the "hits" token bar editable in the FVTT interface.  The system has a "hits" attribute (selectable from the token settings) that is the sum of the damage characteristics (depending on what options are selected).  This release makes the input box corresponding to the bar editable - similar to other systems for FVTT.  So, putting a +6 in the box heals the character by six hp.  Similarly, a -6 causes six damage.  You can also add a total value by entering a value without +/-, e.g. 21, and the damage characteristics totals adjust to achieve that total.  Armor is ignored for these inputs.

Another important note: the order that characteristic damage is added or subtracted is fixed.
- Cepheus Deluxe the order is stamina then lifeblood.
- Cepheus Engine (and similar systems) the order is endurance, strength, then dexterity
- Cepheus Atom it is endurance then strength (as these are proxies for endurance and lifeblood).

You can reverse the order for healing by selecting a new option under ruleset settings: "Reverse characteristic healing order."  This will start healing with the last attribute and work backwards.  For example with CD and the option set to true - the healing order would be lifeblood then stamina.