fix: allow drag drop of lists again (#1157)
This release fixes a feature that broke in transition to v9 - allowing item lists to be re-ordered using drag and drop.  This behavior will override the default of sort alphabetically by item name.  The feature is enabled in the newly renamed "Experimental & Debugging Settings" submenu under "Allow Drag-Drop in Lists".

Also, thanks to @ForjaSalvaje for additional Spanish translations.